Make it easy to use this in cross-compiled projects
This is a very useful crate for `no_std` projects. If we are recompiling, we want to induce cargo to rebuild libcore and cause this crate to use the cross-compiled libcore. By adding this optional dependency, the crate maintains its existing functionality at no extra cost. And those who want to cross-compile can simply use the `cross-compile` feature in their `Cargo.toml`.